refactor(credential): group §12.5 primitives into secrets/ submodule
Moves crypto (AES-256-GCM + Argon2id + PKCE), CredentialGuard,
SecretString, and serde_secret helpers from flat top-level layout
into a secrets/ submodule. Flat root re-exports preserved. The
serde_secret submodule is re-exported at crate root for back-compat
with existing #[serde(with = "nebula_credential::serde_secret")]
attribute paths. Part of phase P2 of credential cleanup.

Also add !crates/credential/src/secrets/ exception to .gitignore so
the new source directory is tracked (sibling pattern to the existing
!crates/credential/src/credentials/ exception).
