+++ crates/credential/src/secrets/guard.rs
