+++ crates/credential/src/secrets/secret_string.rs
