+++ crates/credential/src/secrets/serde_secret.rs
@@ -6,7 +6,7 @@
 
 use serde::{Deserialize, Deserializer, Serializer};
 
-use crate::SecretString;
+use super::SecretString;
 
 /// Serialize the actual secret value (for encrypted-at-rest storage only).
 pub fn serialize<S: Serializer>(secret: &SecretString, s: S) -> Result<S::Ok, S::Error> {
